LayoutTests/storage/indexeddb/open-db-private-browsing-expected.txt

Simple test to open IndexedDB database in private browsing mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

PASS indexedDB.open('db') is null
PASS successfullyParsed is true

TEST COMPLETE
